% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.r
\name{oiler_diagram}
\alias{oiler_diagram}
\title{Solve and draw an Euler screening diagram}
\usage{
oiler_diagram(
  p,
  alpha,
  beta,
  init = c(sqrt(p/(1 - p)), 1),
  draw_pts = 0,
  jitter = 0.1,
  focus = FALSE
)
}
\arguments{
\item{p}{Base rate of positive cases}

\item{alpha}{Probability of true positive, p(T|T)}

\item{beta}{Probability of false positive, p(T|F)}
}
\value{
Solution parameters
}
\description{
This function take the base rate and detection probabilities. It then solves
and draws the corresponding Euler diagram for the implied screening scenario.
}
